requirements.txt
Django==3.2
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.5
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
